Cleanup and return custom AsyncSequence
